Wrote select and indInsertParent
Both are similar functions, but with different uses.
1) Select, finds the node in the tree, and returns an external node if the node doesn't exist.
2) findInsertParent - finds the node underwhich to insert a new node.
The only issue that might arise in insert is if we want to insert a node with a key that is already in the tree.
If we could return a tuple in Java that would solve it (and we can just have 1 function) but I don't know if that is possible